Initial code for the robot is written, to be tested
Changes in robot-config.cpp and main.cpp files, with a lot of comments describing the work added:

Initialized controller and added a code template into the main function, as I am unsure how the userControl() function works for the user control period. This is it for the morning before club meeting. Code is yet to be reviewed and discussed by the other programmers.